# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Products Management",
        "version": "v1"
    },
    "paths": {
...
        "/api/product/": {
            "get": {
                "tags": [
                    "products"
                ],
                "summary": "Find all products",
                "description": "Find all products",
                "operationId": "indexProducts",
                "parameters": [
                    {
                        "name": "Authorization",
                        "in": "header",
                        "description": "Bearer Token",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Not Authorized"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "id": {
                                                "type": "integer",
                                                "example": 1
                                            },
                                            "name": {
                                                "type": "string",
                                                "example": "bananas"
                                            },
                                            "price": {
                                                "type": "number",
                                                "format": "float",
                                                "example": "2000"
                                            },
                                            "stock": {
                                                "type": "integer",
                                                "example": 5
                                            },
                                            "created_at": {
                                                "type": "string",
                                                "example": "2024-05-17T22:04:48.000000Z"
                                            },
                                            "updated_at": {
                                                "type": "string",
                                                "example": "2024-05-17T22:04:48.000000Z"
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "products"
                ],
                "summary": "Create product",
                "description": "Create product",
                "operationId": "storeProduct",
                "parameters": [
                    {
                        "name": "Authorization",
                        "in": "header",
                        "description": "Bearer Token",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "bananas"
                                    },
                                    "price": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 2000
                                    },
                                    "stock": {
                                        "type": "integer",
                                        "example": 5
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Not Authorized"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "201": {
                        "description": "Created",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "name": {
                                            "type": "string",
                                            "example": "bananas"
                                        },
                                        "price": {
                                            "type": "number",
                                            "format": "float",
                                            "example": "2000"
                                        },
                                        "stock": {
                                            "type": "integer",
                                            "example": 5
                                        },
                                        "created_at": {
                                            "type": "string",
                                            "example": "2024-05-17T22:04:48.000000Z"
                                        },
                                        "updated_at": {
                                            "type": "string",
                                            "example": "2024-05-17T22:04:48.000000Z"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/api/product/category/{categoryId}": {
            "get": {
                "tags": [
                    "products"
                ],
                "summary": "Find all products by category id",
                "description": "Find all products by category id",
                "operationId": "getByCategoryIdProduct",
                "parameters": [
                    {
                        "name": "Authorization",
                        "in": "header",
                        "description": "Bearer Token",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "categoryId",
                        "in": "path",
                        "description": "Category id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Not Authorized"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "id": {
                                                "type": "integer",
                                                "example": 1
                                            },
                                            "name": {
                                                "type": "string",
                                                "example": "bananas"
                                            },
                                            "price": {
                                                "type": "number",
                                                "format": "float",
                                                "example": "2000"
                                            },
                                            "stock": {
                                                "type": "integer",
                                                "example": 5
                                            },
                                            "created_at": {
                                                "type": "string",
                                                "example": "2024-05-17T22:04:48.000000Z"
                                            },
                                            "updated_at": {
                                                "type": "string",
                                                "example": "2024-05-17T22:04:48.000000Z"
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
...
    }
}